ques15

+++ ques15.c
@@ -16,7 +16,7 @@ int ans15(int x, int n) {
 	if(x<pow(2,n)){			//if x is less than 2^n, return x
 		return x;
 	}
-	return (x%(pow(2,n)));	//Otherwise return x mod 2^n, the remainder when diving x by 2^n
+	return (x%(int)(pow(2,n)));	//Otherwise return x mod 2^n, the remainder when diving x by 2^n
 }
 
 int main () {
